+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Change Log
 
+## [0.2.0] - **TODO: ADD DATE**
+
+- Pass Props to the root svg element.
+- Allow usage of `comp` instead of `component`.
+
 ## [0.1.2] - 2021-08-07
+
 - fix error ocurred when the component is called without props
 
 ## [0.1.1] - 2021-08-07
+
 - fix passing props to the SVG Component
 
 ## [0.1.0] - 2021-08-07
+
 First release